add produk and produk-detail for admin

+++ adminpanel/kategori-detail.php
@@ -72,6 +72,16 @@ $data = mysqli_fetch_array($query);
         }
 
         if (isset($_POST['deleteBtn'])) {
+            $querycheck = mysqli_query($koneksi, "SELECT * FROM produk WHERE kategori_id = '$id'");
+            $datacount = mysqli_num_rows($querycheck);
+            if ($datacount>0) {
+                ?>
+                <div class="alert alert-warning mt-3" role="alert">
+                            Kategori tidak bisa dihapus karena sudah ada produk
+                        </div>
+                <?php
+                die();
+            }
             $queryDelete = mysqli_query($koneksi, "DELETE FROM kategori WHERE id='$id'");
             if($queryDelete){
                 ?>
